Remove resource definition
